refactor: move CliOptions to separate module for better organization
- Extract CliOptions dataclass from __main__.py to models/cli.py
- Update imports in initialization.py and state.py to use new location
- Remove TYPE_CHECKING import guard as CliOptions is now a proper module
- Clean up empty cli.py file and add proper implementation
- Update test imports to use new CliOptions location
