+++ haproxy_template_ic/models/cli.py
@@ -1,0 +1,9 @@
+from dataclasses import dataclass
+
+
+@dataclass(frozen=True)
+class CliOptions:
+    """Container for bootstrap CLI options (configmap and secret location)."""
+
+    configmap_name: str
+    secret_name: str
